# doc/api/openapi.json
{
  "components": {
...
    "schemas": {
...
      "Deleted": {
        "properties": {
          "deleted": {
            "enum": [
              true
            ],
            "type": "boolean"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "id"
        ],
        "type": "object"
      },
...
      "RunPage": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array"
          },
          "hasMore": {
            "type": "boolean"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "data",
          "nextCursor",
          "hasMore"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "An API key: `Authorization: Bearer sce_live_…`. A browser session token from Clerk is also accepted, which is what the web app's playground uses.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "API support",
      "url": "https://github.com/gitdeepaks/self-consistency-answer-engine"
    },
    "description": "Ask several frontier models the same question, then have an evaluator merge their answers into one — with the agreements, the disagreements and a confidence score attached.\n\n**Authentication.** Every route takes `Authorization: Bearer sce_live_…`. Create a key in the web app under Settings → API keys. A key carries scopes; a request that needs one the key lacks is refused with 403 and `code: \"forbidden\"`.\n\n**Errors.** Every refusal is `{ code, message, details?, requestId }`. Branch on `code`, never on the HTTP status alone — 429 covers both a plan ceiling and a per-window request budget, and `details` tells them apart.\n\n**Retries.** Send an `Idempotency-Key` on every POST. A retry carrying the same key replays the original response instead of performing the write twice.\n\n**Runs are asynchronous.** `POST /v1/runs` returns as soon as the run is queued. Follow it with the SSE stream, or register a webhook and be told.",
    "license": {
      "name": "MIT"
    },
    "title": "Self-Consistency Answer Engine API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/runs": {
      "get": {
        "description": "Newest first, cursor-paginated. Pass the previous page's `nextCursor` back as `cursor`; the value is opaque and must not be constructed by hand.",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "providers",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tags",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "minConfidence",
            "required": false,
            "schema": {
              "maximum": 1,
              "minimum": 0,
              "type": [
                "number",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "mine",
            "required": false,
            "schema": {
              "enum": [
                "true",
                "false"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            },
            "description": "A page of runs."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request did not match the expected shape."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No credential, or one that is expired or revoked."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential lacks the scope or role this route needs."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A plan ceiling or a per-window request budget was reached."
          }
        },
        "summary": "List runs",
        "tags": [
          "Runs"
        ]
      },
      "post": {
        "description": "Fans a prompt out across the model panel and returns immediately. The run is produced asynchronously; follow it with GET /v1/runs/{runId}/events or a run.completed webhook.\n\nSend an Idempotency-Key: a retry carrying the same key returns the original run with 200 rather than fanning out a second, identically expensive panel.",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "maxLength": 255,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "An earlier request with the same Idempotency-Key created this run."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "The run was created."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request did not match the expected shape."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No credential, or one that is expired or revoked."
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The subscription cannot fund new work."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential lacks the scope or role this route needs."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A plan ceiling or a per-window request budget was reached."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The install-wide spend guard is engaged."
          }
        },
        "summary": "Start a run",
        "tags": [
          "Runs"
        ]
      }
    },
    "/runs/{runId}": {
      "delete": {
        "description": "Permanent. Usage records survive the deletion so billing stays correct, but they no longer reference the run.",
        "parameters": [
          {
            "description": "The run's id, as returned by POST /v1/runs.",
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "description": "The run's id, as returned by POST /v1/runs.",
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deleted"
                }
              }
            },
            "description": "The run was deleted."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request did not match the expected shape."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No credential, or one that is expired or revoked."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential lacks the scope or role this route needs."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such run in this workspace."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A plan ceiling or a per-window request budget was reached."
          }
        },
        "summary": "Delete a run",
        "tags": [
          "Runs"
        ]
      },
      "get": {
        "description": "Returns the run with every candidate and the synthesis.\n\nResponses carry an `ETag`. A finished run never changes, so a client polling one should send the tag back as `If-None-Match` and get a 304 with no body.",
        "parameters": [
          {
            "description": "The run's id, as returned by POST /v1/runs.",
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "description": "The run's id, as returned by POST /v1/runs.",
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "The run."
          },
          "304": {
            "description": "The run is unchanged since the supplied ETag."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request did not match the expected shape."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No credential, or one that is expired or revoked."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential lacks the scope or role this route needs."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such run in this workspace."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A plan ceiling or a per-window request budget was reached."
          }
        },
        "summary": "Fetch a run",
        "tags": [
          "Runs"
        ]
      }
    },
...
    "/runs/{runId}/events": {
      "get": {
        "description": "A Server-Sent Events stream of the run's lifecycle. Each frame's `data` is a RunEvent and its `id` is the durable sequence number.\n\nResume after a disconnect by sending the last id you saw as `Last-Event-ID` (EventSource does this by itself) or as `?afterSeq=`. Events with no id are ephemeral — token deltas — and are not replayed.\n\nThe stream closes after `run.completed`, `run.failed` or `run.canceled`. A `ping` event arrives periodically to keep intermediaries from closing an idle connection.",
        "parameters": [
          {
            "description": "The run's id, as returned by POST /v1/runs.",
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "description": "The run's id, as returned by POST /v1/runs.",
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "Resume after this durable sequence number.",
            "in": "query",
            "name": "afterSeq",
            "required": false,
            "schema": {
              "description": "Resume after this durable sequence number.",
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "in": "header",
            "name": "last-event-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The event stream."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such run in this workspace."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Stream a run's progress",
        "tags": [
          "Runs"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "This deployment",
      "url": "/v1"
    }
  ],
...
}